refactoring of Database class

+++ src/main/java/ohtu/data_access/VinkkiDao.java
@@ -19,24 +19,15 @@ public class VinkkiDao implements Dao<Vinkki, Integer> {
         this.database = database;
     }
     
-    private Vinkki createOneFrom(ResultSet rs) throws SQLException {
-        Integer id = rs.getInt("id");
-        String otsikko = rs.getString("otsikko");
-        String tekija = rs.getString("tekija");
-        String kuvaus = rs.getString("kuvaus");
-        String linkki = rs.getString("linkki");
-        String tagit = rs.getString("tagit");
-        Date luettu = rs.getDate("luettu");
-        
-        Vinkki vinkki = new Vinkki(id, otsikko, tekija, kuvaus, linkki, luettu, null);
-        vinkki.setTagit(tagit);
-        return vinkki;
-    }
-
     @Override
     public Vinkki findOne(Integer key) {
         try {
             Connection conn = database.getConnection();
+            
+            if (conn == null) {
+                return null;
+            }
+            
             PreparedStatement stmt = conn.prepareStatement("SELECT * from Vinkki WHERE id=?");
             stmt.setObject(1, key);
 
@@ -62,6 +53,11 @@ public class VinkkiDao implements Dao<Vinkki, Integer> {
         try {
             List<Vinkki> vinkit = new ArrayList<>();
             Connection conn = database.getConnection();
+            
+            if (conn == null) {
+                return null;
+            }
+            
             PreparedStatement stmt = conn.prepareStatement("SELECT * FROM Vinkki");
 
             ResultSet rs = stmt.executeQuery();
@@ -86,6 +82,11 @@ public class VinkkiDao implements Dao<Vinkki, Integer> {
     public void delete(Integer id) {
         try {
             Connection conn = database.getConnection();
+            
+            if (conn == null) {
+                return;
+            }
+            
             PreparedStatement stmt = conn.prepareStatement("DELETE FROM Vinkki WHERE id = ?");
             stmt.setInt(1, id);
             stmt.executeUpdate();
@@ -115,6 +116,11 @@ public class VinkkiDao implements Dao<Vinkki, Integer> {
             }
 
             Connection conn = database.getConnection();
+            
+            if (conn == null) {
+                return;
+            }
+            
             PreparedStatement stmt = conn.prepareStatement("INSERT INTO Vinkki (otsikko, tekija, kuvaus, linkki, tagit, luettu, isbn) VALUES (?,?,?,?,?,?,?)");
             stmt.setString(1, vinkkiOtsikko);
             stmt.setString(2, vinkkiTekija);
@@ -136,6 +142,11 @@ public class VinkkiDao implements Dao<Vinkki, Integer> {
     public void updateWithKey(Integer id) {
         try {
             Connection conn = database.getConnection();
+            
+            if (conn == null) {
+                return;
+            }
+            
             PreparedStatement stmt = conn.prepareStatement("UPDATE Vinkki SET luettu = ? WHERE id = ?");
             stmt.setDate(1, new Date(System.currentTimeMillis()));
             stmt.setInt(2, id);
@@ -152,6 +163,11 @@ public class VinkkiDao implements Dao<Vinkki, Integer> {
     public Vinkki update(Vinkki updatedOne) {
         try {
             Connection conn = database.getConnection();
+            
+            if (conn == null) {
+                return null;
+            }
+            
             PreparedStatement stmt = conn.prepareStatement("UPDATE Vinkki SET otsikko=?, tekija=?, kuvaus=?, linkki=?, tagit=? WHERE id = ?");
             stmt.setString(1, updatedOne.getOtsikko());
             stmt.setString(2, updatedOne.getTekija());
@@ -169,5 +185,19 @@ public class VinkkiDao implements Dao<Vinkki, Integer> {
             System.out.println("ei toimi yhteys databaseen! \n" + ex);
             return null;
         }
+    }
+    
+    private Vinkki createOneFrom(ResultSet rs) throws SQLException {
+        Integer id = rs.getInt("id");
+        String otsikko = rs.getString("otsikko");
+        String tekija = rs.getString("tekija");
+        String kuvaus = rs.getString("kuvaus");
+        String linkki = rs.getString("linkki");
+        String tagit = rs.getString("tagit");
+        Date luettu = rs.getDate("luettu");
+        
+        Vinkki vinkki = new Vinkki(id, otsikko, tekija, kuvaus, linkki, luettu, null);
+        vinkki.setTagit(tagit);
+        return vinkki;
     }
 }
